imp(core): router files renaming + other small improvements
