+++ packages/core/src/index.ts
@@ -1,7 +1,7 @@
 export { EffectFactory } from './effects/effects.factory';
 export { error$ } from './error/error.effect';
 export { HttpError } from './error/error.model';
-export { combineRoutes } from './router/router.factory';
+export { combineRoutes } from './router/router.combiner';
 export * from './effects/effects.interface';
 export * from './router/router.interface';
 export * from './http.interface';
