+++ packages/core/src/router/specs/router.query.factory.spec.ts
@@ -1,7 +1,7 @@
-import { queryParamsFactory } from '../queryParams.factory';
+import { queryParamsFactory } from '../router.query.factory';
 
-describe('queryParamsFactory', () => {
-  test('should return empty object when passed URL with no valid params', () => {
+describe('#queryParamsFactory', () => {
+  test('returns empty object when passed URL with no valid params', () => {
     // given
     const invalidURLS = ['', null, undefined];
 
@@ -14,7 +14,7 @@ describe('queryParamsFactory', () => {
     });
   });
 
-  test('should return empty string when no value for provided key', () => {
+  test('returns empty string when no value for provided key', () => {
     // given
     const exampleURL = 'testParam';
     const expectedParamsObj = { testParam: '' };
@@ -26,7 +26,7 @@ describe('queryParamsFactory', () => {
     expect(params).toEqual(expectedParamsObj);
   });
 
-  test('should return object with multiple keys', () => {
+  test('returns object with multiple keys', () => {
     // given
     const exampleURL = 'testParam=test&testParam2=test2';
     const expectedParamsObj = { testParam: 'test', testParam2: 'test2' };
@@ -38,7 +38,7 @@ describe('queryParamsFactory', () => {
     expect(params).toEqual(expectedParamsObj);
   });
 
-  test('should return array when passed two times same parameter', () => {
+  test('returns array when passed two times same parameter', () => {
     // given
     const exampleURL = 'testParam=test&testParam=123';
     const expectedParamsObj = { testParam: ['test', '123'] };
@@ -50,7 +50,7 @@ describe('queryParamsFactory', () => {
     expect(params).toEqual(expectedParamsObj);
   });
 
-  test('should return nested object', () => {
+  test('returns nested object', () => {
     // given
     const exampleURL = 'testParam[testNested]=test';
     const expectedParamsObj = { testParam: { testNested: 'test' } };
